refactor(config): split server config out to config::server
Server/runtime config is orthogonal to the deployment manifest (it references
no pipeline type), so move it out of the deployment module to its own
`config::server`. Deployment no longer re-exports the server types; call sites
that referenced them through `config::deployment::*` now use `config::server::*`.
`common` primitives stay in `deployment` and are shared by `server`.
