+++ src/config/deployment/tests.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::config::env_var::EnvVarConfig;
 use crate::config::secret_registry::SecretRegistry;
+use crate::config::server::{AllowExecActivities, ServerConfigToml};
 use concepts::cas::Cas;
 use concepts::{ContentDigest, StrVariant, component_id::Digest};
 use sha2::{Digest as _, Sha256};
